+++ app/src/main/AndroidManifest.xml
@@ -9,8 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".CityActivity"></activity>
-        <activity android:name=".ChooseCityActivity">
+        <activity android:name=".CityActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
