+++ app/src/main/res/layout/fragment_choose_city.xml
