+++ app/src/main/res/values/strings.xml
@@ -21,4 +21,5 @@
     <string name="bright">Светлая</string>
     <string name="dark">Темная</string>
     <string name="go_to_wiki">GO TO WIKI</string>
+    <string name="go_back">Назад</string>
 </resources>
